KIEKER-1908 refactored MTree class to be able to better test its parts.
