+++ kieker-analysis/src/kieker/analysis/generic/graph/clustering/MTreeGeneratorStage.java
@@ -50,13 +50,16 @@ public class MTreeGeneratorStage<T> extends AbstractTransformation<List<T>, MTre
 
 		if (models != null) {
 			this.logger.debug("Received {} new models", models.size());
-			for (final T model : models) {
-				mtree.add(model);
+			if (models.size() > 1) {
+				for (final T model : models) {
+					mtree.add(model);
+				}
+				this.logger.debug("Created MTree");
+
+				this.outputPort.send(mtree);
+			} else {
+				this.logger.warn("Received empty model list. Cannot create mtree.");
 			}
-			this.logger.debug("Created MTree");
-
-			this.outputPort.send(mtree);
-
 		}
 	}
