+++ kieker-analysis/src/kieker/analysis/generic/graph/mtree/IRootness.java
@@ -1,0 +1,11 @@
+package kieker.analysis.generic.graph.mtree;
+
+import kieker.analysis.exception.InternalErrorException;
+
+public interface IRootness {
+	int getMinCapacity() throws InternalErrorException;
+
+	void checkDistanceToParent();
+
+	void checkMinCapacity();
+}
