+++ kieker-analysis/src/kieker/analysis/generic/graph/mtree/MTree.java
@@ -15,17 +15,15 @@
  ***************************************************************************/
 package kieker.analysis.generic.graph.mtree;
 
-import java.util.ArrayDeque;
-import java.util.Deque;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.NoSuchElementException;
-import java.util.PriorityQueue;
-import java.util.Set;
-
 import kieker.analysis.exception.InternalErrorException;
-import kieker.analysis.generic.graph.mtree.ISplitFunction.SplitResult;
+import kieker.analysis.generic.graph.mtree.exceptions.DataNotFoundException;
+import kieker.analysis.generic.graph.mtree.exceptions.NodeUnderCapacityException;
+import kieker.analysis.generic.graph.mtree.exceptions.RootNodeReplacementException;
+import kieker.analysis.generic.graph.mtree.exceptions.SplitNodeReplacementException;
+import kieker.analysis.generic.graph.mtree.nodes.AbstractNode;
+import kieker.analysis.generic.graph.mtree.nodes.RootLeafNode;
+import kieker.analysis.generic.graph.mtree.nodes.RootNode;
+import kieker.analysis.generic.graph.mtree.query.Query;
 
 /**
  * The main class that implements the M-Tree.
@@ -50,7 +48,7 @@ public class MTree<T> {
 	protected int maxNodeCapacity;
 	protected IDistanceFunction<? super T> distanceFunction;
 	protected ISplitFunction<T> splitFunction;
-	protected AbstractNode root;
+	protected AbstractNode<T> root;
 
 	/**
 	 * Constructs an M-Tree with the specified distance function.
@@ -82,7 +80,7 @@ public class MTree<T> {
 	public MTree(final int minNodeCapacity,
 			final IDistanceFunction<? super T> distanceFunction,
 			final ISplitFunction<T> splitFunction) {
-		this(minNodeCapacity, 2 * minNodeCapacity - 1, distanceFunction, splitFunction);
+		this(minNodeCapacity, (2 * minNodeCapacity) - 1, distanceFunction, splitFunction);
 	}
 
 	/**
@@ -103,9 +101,9 @@ public class MTree<T> {
 	public MTree(final int minNodeCapacity, final int maxNodeCapacity,
 			final IDistanceFunction<? super T> distanceFunction,
 			final ISplitFunction<T> existingSplitFunction) {
-		if (minNodeCapacity < 2
-				|| maxNodeCapacity <= minNodeCapacity
-				|| distanceFunction == null) {
+		if ((minNodeCapacity < 2)
+				|| (maxNodeCapacity <= minNodeCapacity)
+				|| (distanceFunction == null)) {
 			throw new IllegalArgumentException();
 		}
 
@@ -137,7 +135,7 @@ public class MTree<T> {
 	 */
 	public void add(final T data) throws InternalErrorException {
 		if (this.root == null) {
-			this.root = new RootLeafNode(data);
+			this.root = new RootLeafNode<T>(this, data);
 			try {
 				this.root.addData(data, 0);
 			} catch (final SplitNodeReplacementException e) {
@@ -148,11 +146,11 @@ public class MTree<T> {
 			try {
 				this.root.addData(data, distance);
 			} catch (final SplitNodeReplacementException e) {
-				final AbstractNode newRoot = new RootNode(data);
+				final AbstractNode<T> newRoot = new RootNode<T>(this, data);
 				this.root = newRoot;
-				for (final Object newNode2 : e.newNodes) {
+				for (final Object element : e.getNewNodes()) {
 					@SuppressWarnings("unchecked")
-					final AbstractNode newNode = (AbstractNode) newNode2;
+					final AbstractNode<T> newNode = (AbstractNode<T>) element;
 					distance = this.distanceFunction.calculate(this.root.getData(), newNode.getData());
 					this.root.addChild(newNode, distance);
 				}
@@ -179,7 +177,7 @@ public class MTree<T> {
 			this.root.removeData(data, distanceToRoot);
 		} catch (final RootNodeReplacementException e) {
 			@SuppressWarnings("unchecked")
-			final AbstractNode newRoot = (AbstractNode) e.newRoot;
+			final AbstractNode<T> newRoot = (AbstractNode<T>) e.getNewRoot();
 			this.root = newRoot;
 		} catch (final DataNotFoundException e) {
 			return false;
@@ -199,7 +197,7 @@ public class MTree<T> {
 	 *            neighbors.
 	 * @return A {@link Query} object used to iterate on the results.
 	 */
-	public Query getNearestByRange(final T queryData, final double range) {
+	public Query<T> getNearestByRange(final T queryData, final double range) {
 		return this.getNearest(queryData, range, Integer.MAX_VALUE);
 	}
 
@@ -213,7 +211,7 @@ public class MTree<T> {
 	 *            The maximum number of neighbors to fetch.
 	 * @return A {@link Query} object used to iterate on the results.
 	 */
-	public Query getNearestByLimit(final T queryData, final int limit) {
+	public Query<T> getNearestByLimit(final T queryData, final int limit) {
 		return this.getNearest(queryData, Double.POSITIVE_INFINITY, limit);
 	}
 
@@ -230,8 +228,8 @@ public class MTree<T> {
 	 *            The maximum number of neighbors to fetch.
 	 * @return A {@link Query} object used to iterate on the results.
 	 */
-	public Query getNearest(final T queryData, final double range, final int limit) {
-		return new Query(queryData, range, limit);
+	public Query<T> getNearest(final T queryData, final double range, final int limit) {
+		return new Query<T>(this, queryData, range, limit);
 	}
 
 	/**
@@ -241,8 +239,8 @@ public class MTree<T> {
 	 *            The query data object.
 	 * @return A {@link Query} object used to iterate on the results.
 	 */
-	public Query getNearest(final T queryData) {
-		return new Query(queryData, Double.POSITIVE_INFINITY, Integer.MAX_VALUE);
+	public Query<T> getNearest(final T queryData) {
+		return new Query<T>(this, queryData, Double.POSITIVE_INFINITY, Integer.MAX_VALUE);
 	}
 
 	protected void check() {
@@ -251,866 +249,24 @@ public class MTree<T> {
 		}
 	}
 
-	private class IndexItem { // NOCS cannot be declared final
-		protected double radius;
-		private final T data;
-		private double distanceToParent;
-
-		private IndexItem(final T data) {
-			this.data = data;
-			this.radius = 0;
-			this.distanceToParent = -1;
-		}
-
-		public T getData() {
-			return this.data;
-		}
-
-		public double getDistanceToParent() {
-			return this.distanceToParent;
-		}
-
-		public void setDistanceToParent(final double distance) {
-			this.distanceToParent = distance;
-		}
-
-		int check() {
-			this.checkRadius();
-			this.checkDistanceToParent();
-			return 1;
-		}
-
-		private void checkRadius() {
-			assert this.radius >= 0;
-		}
-
-		protected void checkDistanceToParent() {
-			assert !(this instanceof MTree.RootLeafNode);
-			assert !(this instanceof MTree.RootNode);
-			assert this.distanceToParent >= 0;
-		}
-
+	public int getMaxNodeCapacity() {
+		return this.maxNodeCapacity;
 	}
 
-	private abstract class AbstractNode extends IndexItem {
-
-		protected Map<T, IndexItem> children = new HashMap<>();
-		protected IRootness rootness;
-		protected ILeafness<T> leafness;
-
-		private <R extends AbstractNodeTrait & IRootness, L extends AbstractNodeTrait & ILeafness<T>> AbstractNode(final T data,
-				final R rootness, final L leafness) {
-			super(data);
-
-			rootness.thisNode = this;
-			this.rootness = rootness;
-
-			leafness.thisNode = this;
-			this.leafness = leafness;
-		}
-
-		private final void addData(final T data, final double distance) throws SplitNodeReplacementException, InternalErrorException {
-			this.doAddData(data, distance);
-			this.checkMaxCapacity();
-		}
-
-		@Override
-		int check() {
-			super.check();
-			this.checkMinCapacity();
-			this.checkMaxCapacity2();
-
-			int childHeight = -1;
-			for (final Map.Entry<T, IndexItem> e : this.children.entrySet()) {
-				final T data = e.getKey();
-				final IndexItem child = e.getValue();
-				assert child.data.equals(data);
-
-				this.checkChildClass(child);
-				this.checkChildMetrics(child);
-
-				final int height = child.check();
-				if (childHeight < 0) {
-					childHeight = height;
-				} else {
-					assert childHeight == height;
-				}
-			}
-
-			return childHeight + 1;
-		}
-
-		protected void doAddData(final T data, final double distance) throws InternalErrorException {
-			this.leafness.doAddData(data, distance);
-		}
-
-		protected void doRemoveData(final T data, final double distance) throws DataNotFoundException, NodeUnderCapacityException, InternalErrorException {
-			this.leafness.doRemoveData(data, distance);
-		}
-
-		private final void checkMaxCapacity() throws SplitNodeReplacementException, InternalErrorException {
-			if (this.children.size() > MTree.this.maxNodeCapacity) {
-				final IDistanceFunction<? super T> cachedDistanceFunction = DistanceFunctionFactory.cached(MTree.this.distanceFunction);
-				final SplitResult<T> splitResult = MTree.this.splitFunction.process(this.children.keySet(), cachedDistanceFunction);
-
-				AbstractNode newNode0 = null;
-				AbstractNode newNode1 = null;
-				for (int i = 0; i < 2; ++i) {
-					final T promotedData = splitResult.getPromoted().get(i);
-					final Set<T> partition = splitResult.getPartitions().get(i);
-
-					final AbstractNode newNode = this.newSplitNodeReplacement(promotedData);
-					for (final T data : partition) {
-						final IndexItem child = this.children.get(data);
-						this.children.remove(data);
-						final double distance = cachedDistanceFunction.calculate(promotedData, data);
-						newNode.addChild(child, distance);
-					}
-
-					if (i == 0) {
-						newNode0 = newNode;
-					} else {
-						newNode1 = newNode;
-					}
-				}
-				assert this.children.isEmpty();
-
-				throw new SplitNodeReplacementException(newNode0, newNode1);
-			}
-
-		}
-
-		protected AbstractNode newSplitNodeReplacement(final T data) {
-			return this.leafness.newSplitNodeReplacement(data);
-		}
-
-		protected void addChild(final IndexItem child, final double distance) throws InternalErrorException {
-			this.leafness.addChild(child, distance);
-		}
-
-		void removeData(final T data, final double distance)
-				throws RootNodeReplacementException, NodeUnderCapacityException, DataNotFoundException, InternalErrorException {
-			this.doRemoveData(data, distance);
-			if (this.children.size() < this.getMinCapacity()) {
-				throw new NodeUnderCapacityException();
-			}
-		}
-
-		protected int getMinCapacity() throws InternalErrorException {
-			return this.rootness.getMinCapacity();
-		}
-
-		private void updateMetrics(final IndexItem child, final double distance) {
-			child.setDistanceToParent(distance);
-			this.updateRadius(child);
-		}
-
-		private void updateRadius(final IndexItem child) {
-			this.radius = Math.max(this.radius, child.distanceToParent + child.radius);
-		}
-
-		void checkMinCapacity() {
-			this.rootness.checkMinCapacity();
-		}
-
-		private void checkMaxCapacity2() {
-			assert this.children.size() <= MTree.this.maxNodeCapacity;
-		}
-
-		private void checkChildClass(final IndexItem child) {
-			this.leafness.checkChildClass(child);
-		}
-
-		private void checkChildMetrics(final IndexItem child) {
-			final double dist = MTree.this.distanceFunction.calculate(child.data, this.getData());
-			assert child.distanceToParent == dist;
-
-			final double sum = child.distanceToParent + child.radius;
-			assert sum <= this.radius;
-		}
-
-		@Override
-		protected void checkDistanceToParent() {
-			this.rootness.checkDistanceToParent();
-		}
-
-		private MTree<T> mtree() {
-			return MTree.this;
-		}
+	public int getMinNodeCapacity() {
+		return this.minNodeCapacity;
 	}
 
-	private abstract class AbstractNodeTrait {
-		protected AbstractNode thisNode;
+	public ISplitFunction<T> getSplitFunction() {
+		return this.splitFunction;
 	}
 
-	private interface ILeafness<DATA> {
-		void doAddData(DATA data, double distance) throws InternalErrorException;
-
-		void addChild(MTree<DATA>.IndexItem child, double distance) throws InternalErrorException;
-
-		void doRemoveData(DATA data, double distance) throws DataNotFoundException, NodeUnderCapacityException, InternalErrorException;
-
-		MTree<DATA>.AbstractNode newSplitNodeReplacement(DATA data);
-
-		void checkChildClass(MTree<DATA>.IndexItem child);
+	public IDistanceFunction<? super T> getDistanceFunction() {
+		return this.distanceFunction;
 	}
 
-	private interface IRootness {
-		int getMinCapacity() throws InternalErrorException;
-
-		void checkDistanceToParent();
-
-		void checkMinCapacity();
-	}
-
-	private class RootNodeTrait extends AbstractNodeTrait implements IRootness {
-
-		public RootNodeTrait() {
-			// default constructor
-		}
-
-		@Override
-		public int getMinCapacity() throws InternalErrorException {
-			throw new InternalErrorException("Should not be called!");
-		}
-
-		@Override
-		public void checkDistanceToParent() {
-			assert this.thisNode.getDistanceToParent() == -1;
-		}
-
-		@Override
-		public void checkMinCapacity() {
-			this.thisNode.checkMinCapacity();
-		}
-
-	};
-
-	private class NonRootNodeTrait extends AbstractNodeTrait implements IRootness {
-
-		public NonRootNodeTrait() {
-			// default constructor
-		}
-
-		@Override
-		public int getMinCapacity() {
-			return MTree.this.minNodeCapacity;
-		}
-
-		@Override
-		public void checkMinCapacity() {
-			assert this.thisNode.children.size() >= this.thisNode.mtree().minNodeCapacity;
-		}
-
-		@Override
-		public void checkDistanceToParent() {
-			assert this.thisNode.getDistanceToParent() >= 0;
-		}
-	};
-
-	private class LeafNodeTrait extends AbstractNodeTrait implements ILeafness<T> {
-
-		public LeafNodeTrait() {
-			// default constructor
-		}
-
-		@Override
-		public void doAddData(final T data, final double distance) {
-			final Entry entry = this.thisNode.mtree().new Entry(data);
-			assert !this.thisNode.children.containsKey(data);
-			this.thisNode.children.put(data, entry);
-			assert this.thisNode.children.containsKey(data);
-			this.thisNode.updateMetrics(entry, distance);
-		}
-
-		@Override
-		public void addChild(final IndexItem child, final double distance) {
-			assert !this.thisNode.children.containsKey(child.data);
-			this.thisNode.children.put(child.data, child);
-			assert this.thisNode.children.containsKey(child.data);
-			this.thisNode.updateMetrics(child, distance);
-		}
-
-		@Override
-		public AbstractNode newSplitNodeReplacement(final T data) {
-			return this.thisNode.mtree().new LeafNode(data);
-		}
-
-		@Override
-		public void doRemoveData(final T data, final double distance) throws DataNotFoundException {
-			if (this.thisNode.children.remove(data) == null) {
-				throw new DataNotFoundException();
-			}
-		}
-
-		@Override
-		public void checkChildClass(final IndexItem child) {
-			assert child instanceof MTree.Entry;
-		}
-	}
-
-	class NonLeafNodeTrait extends AbstractNodeTrait implements ILeafness<T> {
-
-		public NonLeafNodeTrait() {
-			// nothing to be done
-		}
-
-		@Override
-		public void doAddData(final T data, final double distance) throws InternalErrorException {
-			final class CandidateChild {
-				private final AbstractNode node;
-				private final double distance;
-				private final double metric;
-
-				private CandidateChild(final AbstractNode node, final double distance, final double metric) {
-					this.node = node;
-					this.distance = distance;
-					this.metric = metric;
-				}
-
-				public AbstractNode getNode() {
-					return this.node;
-				}
-
-				public double getDistance() {
-					return this.distance;
-				}
-
-				public double getMetric() {
-					return this.metric;
-				}
-			}
-
-			CandidateChild minRadiusIncreaseNeeded = new CandidateChild(null, -1.0, Double.POSITIVE_INFINITY);
-			CandidateChild nearestDistance = new CandidateChild(null, -1.0, Double.POSITIVE_INFINITY);
-
-			for (final IndexItem item : this.thisNode.children.values()) {
-				@SuppressWarnings("unchecked")
-				final AbstractNode child = (AbstractNode) item;
-				final double childDistance = this.thisNode.mtree().distanceFunction.calculate(child.getData(), data);
-				if (childDistance > child.radius) {
-					final double radiusIncrease = childDistance - child.radius;
-					if (radiusIncrease < minRadiusIncreaseNeeded.metric) {
-						minRadiusIncreaseNeeded = new CandidateChild(child, childDistance, radiusIncrease);
-					}
-				} else {
-					if (childDistance < nearestDistance.metric) {
-						nearestDistance = new CandidateChild(child, childDistance, childDistance);
-					}
-				}
-			}
-
-			final CandidateChild chosen = nearestDistance.node != null ? nearestDistance : minRadiusIncreaseNeeded; // NOCS
-
-			final AbstractNode child = chosen.node;
-			try {
-				child.addData(data, chosen.distance);
-				this.thisNode.updateRadius(child);
-			} catch (final SplitNodeReplacementException e) {
-				// Replace current child with new nodes
-				final IndexItem itemIndex = this.thisNode.children.remove(child.getData());
-				assert itemIndex != null;
-
-				for (final Object newNode : e.newNodes) {
-					@SuppressWarnings("unchecked")
-					final AbstractNode newChild = (AbstractNode) newNode;
-					final double newDistance = this.thisNode.mtree().distanceFunction.calculate(this.thisNode.getData(), newChild.getData());
-					this.thisNode.addChild(newChild, newDistance);
-				}
-			}
-		}
-
-		@Override
-		public void addChild(final IndexItem inputNewChild, final double inputDistance) throws InternalErrorException {
-			double distance = inputDistance;
-			@SuppressWarnings("unchecked")
-			AbstractNode newChild = (AbstractNode) inputNewChild;
-
-			final class ChildWithDistance {
-				private final AbstractNode child;
-				private final double distance;
-
-				private ChildWithDistance(final AbstractNode child, final double distance) {
-					this.child = child;
-					this.distance = distance;
-				}
-
-				public AbstractNode getChild() {
-					return this.child;
-				}
-
-				public double getDistance() {
-					return this.distance;
-				}
-			}
-
-			final Deque<ChildWithDistance> newChildren = new ArrayDeque<>();
-			newChildren.addFirst(new ChildWithDistance(newChild, distance));
-
-			while (!newChildren.isEmpty()) {
-				final ChildWithDistance cwd = newChildren.removeFirst();
-
-				newChild = cwd.child;
-				distance = cwd.distance;
-				if (this.thisNode.children.containsKey(newChild.getData())) {
-					@SuppressWarnings("unchecked")
-					final AbstractNode existingChild = (AbstractNode) this.thisNode.children.get(newChild.getData());
-					assert existingChild.getData().equals(newChild.getData());
-
-					// Transfer the _children_ of the newChild to the existingChild
-					for (final IndexItem grandchild : newChild.children.values()) {
-						existingChild.addChild(grandchild, grandchild.distanceToParent);
-					}
-					newChild.children.clear();
-
-					try {
-						existingChild.checkMaxCapacity();
-					} catch (final SplitNodeReplacementException e) {
-						final IndexItem indexItem = this.thisNode.children.remove(existingChild.getData());
-						assert indexItem != null;
-
-						for (final Object newNode2 : e.newNodes) {
-							@SuppressWarnings("unchecked")
-							final AbstractNode newNode = (AbstractNode) newNode2;
-							final double newDistance = this.thisNode.mtree().distanceFunction.calculate(this.thisNode.getData(),
-									newNode.getData());
-							newChildren.addFirst(new ChildWithDistance(newNode, newDistance));
-						}
-					}
-				} else {
-					this.thisNode.children.put(newChild.getData(), newChild);
-					this.thisNode.updateMetrics(newChild, distance);
-				}
-			}
-		}
-
-		@Override
-		public AbstractNode newSplitNodeReplacement(final T data) {
-			return new InternalNode(data);
-		}
-
-		@Override
-		public void doRemoveData(final T data, final double distance) throws DataNotFoundException,
-				NodeUnderCapacityException, InternalErrorException {
-			for (final IndexItem childItem : this.thisNode.children.values()) {
-				@SuppressWarnings("unchecked")
-				final AbstractNode child = (AbstractNode) childItem;
-				if (Math.abs(distance - child.getDistanceToParent()) <= child.radius) {
-					final double distanceToChild = this.thisNode.mtree().distanceFunction.calculate(data, child.getData());
-					if (distanceToChild <= child.radius) {
-						try {
-							child.removeData(data, distanceToChild);
-							this.thisNode.updateRadius(child);
-							return;
-						} catch (final DataNotFoundException e) {
-							// If DataNotFound was thrown, then the data was not found in the child
-						} catch (final NodeUnderCapacityException e) {
-							final AbstractNode expandedChild = this.balanceChildren(child);
-							this.thisNode.updateRadius(expandedChild);
-							return;
-						} catch (final RootNodeReplacementException e) {
-							throw new InternalErrorException("Should never happen!");
-						}
-					}
-				}
-			}
-
-			throw new DataNotFoundException();
-		}
-
-		private AbstractNode balanceChildren(final AbstractNode theChild) throws InternalErrorException {
-			// Tries to find anotherChild which can donate a grand-child to theChild.
-
-			AbstractNode nearestDonor = null;
-			double distanceNearestDonor = Double.POSITIVE_INFINITY;
-
-			AbstractNode nearestMergeCandidate = null;
-			double distanceNearestMergeCandidate = Double.POSITIVE_INFINITY;
-
-			for (final IndexItem child : this.thisNode.children.values()) {
-				@SuppressWarnings("unchecked")
-				final AbstractNode anotherChild = (AbstractNode) child;
-				if (anotherChild == theChild) {
-					continue;
-				}
-
-				final double distance = this.thisNode.mtree().distanceFunction.calculate(theChild.getData(), anotherChild.getData());
-				if (anotherChild.children.size() > anotherChild.getMinCapacity()) {
-					if (distance < distanceNearestDonor) {
-						distanceNearestDonor = distance;
-						nearestDonor = anotherChild;
-					}
-				} else {
-					if (distance < distanceNearestMergeCandidate) {
-						distanceNearestMergeCandidate = distance;
-						nearestMergeCandidate = anotherChild;
-					}
-				}
-			}
-
-			if (nearestDonor == null) {
-				// Merge
-				for (final IndexItem grandchild : theChild.children.values()) {
-					final double distance = this.thisNode.mtree().distanceFunction.calculate(grandchild.getData(),
-							nearestMergeCandidate.getData());
-					nearestMergeCandidate.addChild(grandchild, distance);
-				}
-
-				final IndexItem removed = this.thisNode.children.remove(theChild.getData());
-				assert removed != null;
-				return nearestMergeCandidate;
-			} else {
-				// Donate
-				// Look for the nearest grandchild
-				IndexItem nearestGrandchild = null;
-				double nearestGrandchildDistance = Double.POSITIVE_INFINITY;
-				for (final IndexItem grandchild : nearestDonor.children.values()) {
-					final double distance = this.thisNode.mtree().distanceFunction.calculate(grandchild.data, theChild.getData());
-					if (distance < nearestGrandchildDistance) {
-						nearestGrandchildDistance = distance;
-						nearestGrandchild = grandchild;
-					}
-				}
-
-				final IndexItem indexItem = nearestDonor.children.remove(nearestGrandchild.data);
-				assert indexItem != null;
-				theChild.addChild(nearestGrandchild, nearestGrandchildDistance);
-				return theChild;
-			}
-		}
-
-		@Override
-		public void checkChildClass(final IndexItem child) {
-			assert child instanceof MTree.InternalNode
-					|| child instanceof MTree.LeafNode;
-		}
-	}
-
-	private final class RootLeafNode extends AbstractNode {
-
-		private RootLeafNode(final T data) {
-			super(data, new RootNodeTrait(), new LeafNodeTrait());
-		}
-
-		@Override
-		void removeData(final T data, final double distance) throws RootNodeReplacementException, DataNotFoundException, InternalErrorException {
-			try {
-				super.removeData(data, distance);
-			} catch (final NodeUnderCapacityException e) {
-				assert this.children.isEmpty();
-				throw new RootNodeReplacementException(null);
-			}
-		}
-
-		@Override
-		protected int getMinCapacity() {
-			return 1;
-		}
-
-		@Override
-		void checkMinCapacity() {
-			assert this.children.size() >= 1;
-		}
-	}
-
-	private final class RootNode extends AbstractNode {
-
-		private RootNode(final T data) {
-			super(data, new RootNodeTrait(), new NonLeafNodeTrait());
-		}
-
-		@Override
-		void removeData(final T data, final double distance)
-				throws RootNodeReplacementException, NodeUnderCapacityException, DataNotFoundException, InternalErrorException {
-			try {
-				super.removeData(data, distance);
-			} catch (final NodeUnderCapacityException e) {
-				// Promote the only child to root
-				@SuppressWarnings("unchecked")
-				final AbstractNode theChild = (AbstractNode) this.children.values().iterator().next();
-				final AbstractNode newRoot;
-				if (theChild instanceof MTree.InternalNode) {
-					newRoot = new RootNode(theChild.getData());
-				} else {
-					assert theChild instanceof MTree.LeafNode;
-					newRoot = new RootLeafNode(theChild.getData());
-				}
-
-				for (final IndexItem grandchild : theChild.children.values()) {
-					final double newDistance = MTree.this.distanceFunction.calculate(newRoot.getData(), grandchild.data);
-					newRoot.addChild(grandchild, newDistance);
-				}
-				theChild.children.clear();
-
-				throw new RootNodeReplacementException(newRoot);
-			}
-		}
-
-		@Override
-		protected int getMinCapacity() {
-			return 2;
-		}
-
-		@Override
-		void checkMinCapacity() {
-			assert this.children.size() >= 2;
-		}
-	}
-
-	private final class InternalNode extends AbstractNode {
-		private InternalNode(final T data) {
-			super(data, new NonRootNodeTrait(), new NonLeafNodeTrait());
-		}
-	};
-
-	private class LeafNode extends AbstractNode {
-
-		public LeafNode(final T data) {
-			super(data, new NonRootNodeTrait(), new LeafNodeTrait());
-		}
-	}
-
-	private final class Entry extends IndexItem {
-		private Entry(final T data) {
-			super(data);
-		}
-	}
-
-	/**
-	 * The type of the results for nearest-neighbor queries.
-	 */
-	public final class ResultItem {
-
-		/** A nearest-neighbor. */
-		private final T data;
-
-		/**
-		 * The distance from the nearest-neighbor to the query data object
-		 * parameter.
-		 */
-		private final double distance;
-
-		private ResultItem(final T data, final double distance) {
-			this.data = data;
-			System.err.println("ResultItem data " + data + " " + distance);
-			this.distance = distance;
-		}
-
-		public T getData() {
-			return this.data;
-		}
-
-		public double getDistance() {
-			return this.distance;
-		}
-	}
-
-	// Exception classes
-	private static final class SplitNodeReplacementException extends Exception {
-		// A subclass of Throwable cannot be generic. :-(
-		// So, we have newNodes declared as Object[] instead of Node[].
-		private final Object[] newNodes;
-
-		private SplitNodeReplacementException(final Object... newNodes) {
-			this.newNodes = newNodes;
-		}
-	}
-
-	private static final class RootNodeReplacementException extends Exception {
-		// A subclass of Throwable cannot be generic. :-(
-		// So, we have newRoot declared as Object instead of Node.
-		private final Object newRoot;
-
-		private RootNodeReplacementException(final Object newRoot) {
-			this.newRoot = newRoot;
-		}
-	}
-
-	private static final class NodeUnderCapacityException extends Exception {
-		public NodeUnderCapacityException() {
-			// default constructor
-		}
-	}
-
-	private static final class DataNotFoundException extends Exception {
-		public DataNotFoundException() {
-			// default constructor
-		}
-	}
-
-	/**
-	 * An {@link Iterable} class which can be iterated to fetch the results of a
-	 * nearest-neighbors query.
-	 *
-	 * <p>
-	 * The neighbors are presented in non-decreasing order from the {@code
-	 * queryData} argument to the {@link MTree#getNearest(Object, double, int)
-	 * getNearest*()}
-	 * call.
-	 *
-	 * <p>
-	 * The query on the M-Tree is executed during the iteration, as the
-	 * results are fetched. It means that, by the time when the <i>n</i>-th
-	 * result is fetched, the next result may still not be known, and the
-	 * resources allocated were only the necessary to identify the <i>n</i>
-	 * first results.
-	 */
-	public final class Query implements Iterable<ResultItem> {
-
-		private final T data;
-		private final double range;
-		private final int limit;
-
-		private Query(final T data, final double range, final int limit) {
-			this.data = data;
-			this.range = range;
-			this.limit = limit;
-		}
-
-		@Override
-		public Iterator<ResultItem> iterator() {
-			return new ResultsIterator();
-		}
-
-		private final class ResultsIterator implements Iterator<ResultItem> {
-
-			private ResultItem nextResultItem = null;
-			private boolean finished = false;
-			private final PriorityQueue<ItemWithDistances<AbstractNode>> pendingQueue = new PriorityQueue<>();
-			private double nextPendingMinDistance;
-			private final PriorityQueue<ItemWithDistances<Entry>> nearestQueue = new PriorityQueue<>();
-			private int yieldedCount;
-
-			private ResultsIterator() {
-				if (MTree.this.root == null) {
-					this.finished = true;
-					return;
-				}
-
-				final double distance = MTree.this.distanceFunction.calculate(Query.this.data, MTree.this.root.getData());
-				final double minDistance = Math.max(distance - MTree.this.root.radius, 0.0);
-
-				this.pendingQueue.add(new ItemWithDistances<>(MTree.this.root, distance, minDistance));
-				this.nextPendingMinDistance = minDistance;
-			}
-
-			@Override
-			public boolean hasNext() {
-				if (this.finished) {
-					return false;
-				}
-
-				if (this.nextResultItem == null) {
-					this.fetchNext();
-				}
-
-				if (this.nextResultItem == null) {
-					this.finished = true;
-					return false;
-				} else {
-					return true;
-				}
-			}
-
-			@Override
-			public ResultItem next() {
-				if (this.hasNext()) {
-					final ResultItem next = this.nextResultItem;
-					this.nextResultItem = null;
-					return next;
-				} else {
-					throw new NoSuchElementException();
-				}
-			}
-
-			@Override
-			public void remove() {
-				throw new UnsupportedOperationException();
-			}
-
-			private void fetchNext() {
-				assert !this.finished;
-
-				if (this.finished || this.yieldedCount >= Query.this.limit) {
-					this.finished = true;
-					return;
-				}
-
-				while (!this.pendingQueue.isEmpty() || !this.nearestQueue.isEmpty()) {
-					if (this.prepareNextNearest()) {
-						return;
-					}
-
-					assert !this.pendingQueue.isEmpty();
-
-					final ItemWithDistances<AbstractNode> pending = this.pendingQueue.poll();
-					final AbstractNode node = pending.item;
-
-					for (final IndexItem child : node.children.values()) {
-						if (Math.abs(pending.distance - child.distanceToParent) - child.radius <= Query.this.range) {
-							final double childDistance = MTree.this.distanceFunction.calculate(Query.this.data, child.data);
-							final double childMinDistance = Math.max(childDistance - child.radius, 0.0);
-							if (childMinDistance <= Query.this.range) {
-								if (child instanceof MTree.Entry) {
-									@SuppressWarnings("unchecked")
-									final Entry entry = (Entry) child;
-									this.nearestQueue.add(new ItemWithDistances<>(entry, childDistance, childMinDistance));
-								} else {
-									@SuppressWarnings("unchecked")
-									final AbstractNode childNode = (AbstractNode) child;
-									this.pendingQueue.add(new ItemWithDistances<>(childNode, childDistance, childMinDistance));
-								}
-							}
-						}
-					}
-
-					if (this.pendingQueue.isEmpty()) {
-						this.nextPendingMinDistance = Double.POSITIVE_INFINITY;
-					} else {
-						this.nextPendingMinDistance = this.pendingQueue.peek().minDistance;
-					}
-				}
-
-				this.finished = true;
-			}
-
-			private boolean prepareNextNearest() {
-				if (!this.nearestQueue.isEmpty()) {
-					final ItemWithDistances<Entry> nextNearest = this.nearestQueue.peek();
-					if (nextNearest.distance <= this.nextPendingMinDistance) {
-						this.nearestQueue.poll();
-						this.nextResultItem = new ResultItem(nextNearest.item.getData(), nextNearest.distance);
-						++this.yieldedCount;
-						return true;
-					}
-				}
-
-				return false;
-			}
-
-			private class ItemWithDistances<U> implements Comparable<ItemWithDistances<U>> {
-				private final U item;
-				private final double distance;
-				private final double minDistance;
-
-				public ItemWithDistances(final U item, final double distance, final double minDistance) {
-					this.item = item;
-					this.distance = distance;
-					this.minDistance = minDistance;
-				}
-
-				@Override
-				public int compareTo(final ItemWithDistances<U> that) {
-					if (this.minDistance < that.minDistance) {
-						return -1;
-					} else if (this.minDistance > that.minDistance) {
-						return +1;
-					} else {
-						return 0;
-					}
-				}
-			}
-		}
+	public AbstractNode<T> getRoot() {
+		return this.root;
 	}
 
 }
