+++ kieker-analysis/src/kieker/analysis/generic/graph/mtree/exceptions/DataNotFoundException.java
@@ -1,0 +1,10 @@
+package kieker.analysis.generic.graph.mtree.exceptions;
+
+public final class DataNotFoundException extends Exception {
+
+	private static final long serialVersionUID = 6075822198494027074L;
+
+	public DataNotFoundException() {
+		// default constructor
+	}
+}
