+++ kieker-analysis/src/kieker/analysis/generic/graph/mtree/exceptions/NodeUnderCapacityException.java
@@ -1,0 +1,7 @@
+package kieker.analysis.generic.graph.mtree.exceptions;
+
+public final class NodeUnderCapacityException extends Exception {
+	public NodeUnderCapacityException() {
+		// default constructor
+	}
+}
