+++ kieker-analysis/src/kieker/analysis/generic/graph/mtree/nodes/AbstractNodeTrait.java
@@ -1,0 +1,5 @@
+package kieker.analysis.generic.graph.mtree.nodes;
+
+public abstract class AbstractNodeTrait<T> {
+	protected AbstractNode<T> thisNode;
+}
