+++ kieker-analysis/src/kieker/analysis/generic/graph/mtree/nodes/Entry.java
@@ -1,0 +1,7 @@
+package kieker.analysis.generic.graph.mtree.nodes;
+
+public final class Entry<T> extends IndexItem<T> {
+	public Entry(final T data) {
+		super(data);
+	}
+}
