+++ kieker-analysis/src/kieker/analysis/generic/graph/mtree/nodes/NonRootNodeTrait.java
@@ -1,0 +1,25 @@
+package kieker.analysis.generic.graph.mtree.nodes;
+
+import kieker.analysis.generic.graph.mtree.IRootness;
+
+public class NonRootNodeTrait<T> extends AbstractNodeTrait<T> implements IRootness {
+
+	public NonRootNodeTrait() {
+		super();
+	}
+
+	@Override
+	public int getMinCapacity() {
+		return this.thisNode.getMTree().getMinNodeCapacity();
+	}
+
+	@Override
+	public void checkMinCapacity() {
+		assert this.thisNode.children.size() >= this.thisNode.getMTree().getMinNodeCapacity();
+	}
+
+	@Override
+	public void checkDistanceToParent() {
+		assert this.thisNode.getDistanceToParent() >= 0;
+	}
+};
