+++ kieker-analysis/src/kieker/analysis/generic/graph/mtree/query/package-info.java
@@ -1,0 +1,1 @@
+package kieker.analysis.generic.graph.mtree.query;
